Give the two bulk email modules a test at their mirror path
The mutation gate refuses a changed source with no direct test at its
mirror path, so this branch could not run it. Neither module had one.

`test/features/admin/bulk-email.test.ts` is the former
`test/integration/server/bulk-email/send.test.ts`, moved unchanged. It
drives the admin routes exactly like its 313 siblings under
`test/features/`.

`test/shared/bulk-email.test.ts` holds the `summarizeProviderResponse`
tests, moved out of `test/integration/bulk-email.test.ts`. That file was
597 lines, over the size this repository aims for, and the block that
moved is the part this branch changed.
